+++ src/Service/Menu/MenuBuilder.php
@@ -45,12 +45,12 @@ readonly class MenuBuilder
 
         $superAdminToolsSubItem = (new MenuItem(label: 'Outils SA', roleGranted: User::ROLE_ADMIN))
             ->addChild(new MenuItem(label: 'Partenaires archivés', route: 'back_archived_partner_index', roleGranted: User::ROLE_ADMIN))
-            ->addChild(new MenuItem(label: 'Comptes archivés', route: 'back_account_index', roleGranted: User::ROLE_ADMIN))
+            ->addChild(new MenuItem(label: 'Comptes archivés', route: 'back_archived_users_index', roleGranted: User::ROLE_ADMIN))
             ->addChild(new MenuItem(label: 'Signalement archivés', route: 'back_archived_signalements_index', roleGranted: User::ROLE_ADMIN))
             ->addChild(new MenuItem(label: 'Règles d\'auto-affectation', route: 'back_auto_affectation_rule_index', roleGranted: User::ROLE_ADMIN))
             ->addChild(new MenuItem(label: 'Résumés de suivis', route: 'back_suivi_summaries_index', roleGranted: User::ROLE_ADMIN))
             ->addChild(new MenuItem(label: 'Territoires', route: 'back_territory_index', roleGranted: User::ROLE_ADMIN))
-            ->addChild(new MenuItem(route: 'back_account_reactiver'))
+            ->addChild(new MenuItem(route: 'back_archived_users_reactiver'))
             ->addChild(new MenuItem(route: 'back_territory_edit'))
             ->addChild(new MenuItem(route: 'back_auto_affectation_rule_new'))
             ->addChild(new MenuItem(route: 'back_auto_affectation_rule_edit'))
