+++ templates/back/account/_form.html.twig
@@ -49,7 +49,7 @@
 <div class="fr-grid-row">
     <div class="fr-col-6">
         <a class="fr-btn fr-btn--secondary fr-fi-arrow-left-line fr-btn--icon-left"
-        href="{{ path('back_account_index') }}"
+        href="{{ path('back_archived_users_index') }}"
         data-filter-list-comptes-archives>Annuler</a>
     </div>
     <div class="fr-col-6 fr-text--right">
