+++ templates/back/account/edit.html.twig
@@ -9,7 +9,7 @@
                 'level2Link': '',
                 'level2Label': '',
                 'level3Title': 'Comptes archivés',
-                'level3Link': path('back_account_index'),
+                'level3Link': path('back_archived_users_index'),
                 'level3Label': 'Retour à la liste des comptes archivés',
                 'level4Title': 'Réactivation du compte',
                 'level4Link': '',
